Update to ember-remarkable 3.6.0
NOTE: Currently on a fork until
https://github.com/johnotander/ember-remarkable/pull/34 is merged

This commit pulls the injection of the "hljs" class selector into a
freestyle-md-text component that extends md-text from ember-remarkable.
Prior to this, this hack had been done on an ember-remarkable fork I
was maintaining. It's nice to be rid of that bit of silliness.