Delete environment.js
This is now merged into the app environment and pollutes the rooURL parameter in production builds.

The original reason for this configuration is to support deploying freestyle to github pages. This should still work and is controlled by the dummy app `config/environment` file, which already has the same information this file has. 

The reason this is breaking now is due to updates in how ember-cli merges in environments from addons in 3.14...I think.