Move styles into app namespace to avoid ember-cli-sass addon.scss error
The ember-cli-sass plugin has a bug that throws an error about a missing
addon.sass file of the plugin when
* The host app and one of it's addons both use ember-cli-sass addon and
* The host app uses .sass extension and
* The plugin uses .scss extension and
* The plugin has a addon/styles directory

All 4 Conditions are required to reproduce. Best practice amongst ember-
cli-sass addon authors became to add stylesheets to the apps namespace
and require the developer to import these explicitly into the app's
styles. This solution allows both the app and the addon author to use
sass with whatever extension they prefer.