chore!: remove the use of `ember-json-viewer`
`ember-json-viewer` is still required to offer support for Ember v2.12.
We've tried updating the addon to provide support for Ember v4 while
still maintaining support for v2.12, but this wasn't very straight forward.
For the time being, it's easier to just drop `ember-json-viewer` as a dependency
and use an alternative instead, being `json-formatter-js`.