Rework the @cssVariable decorator
Typically, component-specific CSS variables will be initialized in the
base component selector. The class name of the component can now be
specified using the cssClassName decorator option or
cssClassName class variable.

The decorator now also accepts a toVariableName callback, to allow
specifying how to convert the decorated property name to a CSS
variable name. The default implementation dasherizes it.